Despaghettify 4 Moved debug drawing stuff into separate object
