+++ src/main/java/de/officeryoda/fhysics/engine/FhysicsCore.kt
@@ -1,8 +1,6 @@
 package de.officeryoda.fhysics.engine
 
-import de.officeryoda.fhysics.engine.objects.Circle
 import de.officeryoda.fhysics.engine.objects.FhysicsObject
-import de.officeryoda.fhysics.engine.objects.FhysicsObjectFactory
 import de.officeryoda.fhysics.extensions.times
 import de.officeryoda.fhysics.rendering.FhysicsObjectDrawer
 import de.officeryoda.fhysics.rendering.GravityType
@@ -43,11 +41,11 @@ object FhysicsCore {
 
     init {
 
-        for (i in 1..3000) {
-            val circle: Circle = FhysicsObjectFactory.randomCircle()
-//            circle.velocity.set(Vector2.ZERO)
-            spawn(circle)
-        }
+//        for (i in 1..3000) {
+//            val circle: Circle = FhysicsObjectFactory.randomCircle()
+////            circle.velocity.set(Vector2.ZERO)
+//            spawn(circle)
+//        }
 //
 //        for (i in 1..20) {
 //            val rect: Rectangle = FhysicsObjectFactory.randomRectangle()
